// src/generated/graphql.ts
export enum Mastery {
    Advanced = "Advanced",
    Competent = "Competent",
    Expert = "Expert",
    Novice = "Novice",
    Proficient = "Proficient",
}
// ...
export enum Proficiency {
    A1 = "A1",
    A2 = "A2",
    B1 = "B1",
    B2 = "B2",
    C1 = "C1",
    C2 = "C2",
    Native = "Native",
}